Modernized the Makefiles for the standard targets: all clean docs includes install libs test
Gnu Make is now required.

Make can be invoked using three use cases:  Developer working
in their sandbox, and two multi-root modes (pwd is top of source
tree, or pwd is top of build tree, where source and build trees
may be rooted in different directories).

Developers can simple make targets directly in their sandboxes
(e.g. make all) when the pwd is one of the component source
directories.

If the pwd is the top of the build tree, Make can be invoked with
the -f option to specify the Makefile in the top of the source tree.
The derived files will be written to the build tree.

If the pwd is the top of the source tree, Make can be invoked by
setting the BUILDROOT macro to the fullpath of the root of the build
tree.  This can be done on the Make command line.  The derived files
will be written to the build tree.  By default, both the source tree
and the build tree are rooted at the directory containing the
top-level Makefile.

The PREFIX macro specifies the location of the final installation.
Its default value is the /usr/local directory.  It can be set on
the Make command line.

The following standard targets are defined at the top level and
in each component:

all: Same as building the docs, includes, and libs targets.
clean: Deletes derived files from the build tree.
docs: Copies man pages to the build tree.
includes: Copies header files to the build tree.
install: Copies installable files to the specified PREFIX directory tree.
libs: Compiles the libchipset.a library.
test: Runs unit tests and reports results.